fix DrawerTriggerButton component tests
